src/requirements.txt
ruamel.yaml
docker
GitPython
setuptools